# rust/tic_tac_toe_server_rust/openapi.json
{"openapi":"3.1.0","info":{"title":"tic_tac_toe_server_rust","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/available_agents":{"get":{"tags":["Tic Tac Toe"],"operationId":"get_available_agents_handler","responses":{"200":{"description":"利用可能なエージェントのリスト","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailableAgentsResponse"}}}}}}},"/game_status":{"get":{"tags":["Tic Tac Toe"],"operationId":"get_game_status_handler","responses":{"200":{"description":"現在のゲームの状態を取得","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardState"}}}},"404":{"description":"ゲームが開始されていません","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/make_move":{"post":{"tags":["Tic Tac Toe"],"operationId":"make_move_handler","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoveRequest"}}},"required":true},"responses":{"200":{"description":"移動が正常に実行され、最新のゲーム状態が返されました","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardState"}}}},"400":{"description":"不正な移動です","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"サーバーエラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/start_game":{"post":{"tags":["Tic Tac Toe"],"operationId":"start_game_handler","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartGameRequest"}}},"required":true},"responses":{"200":{"description":"新しいゲームが開始されました","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BoardState"}}}},"500":{"description":"ゲームの開始に失敗しました","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"AvailableAgentsResponse":{"type":"object","required":["agents"],"properties":{"agents":{"type":"array","items":{"type":"string"}}}},"BoardState":{"type":"object","required":["board","current_player","game_over"],"properties":{"board":{"type":"array","items":{"type":"array","items":{"$ref":"#/components/schemas/Player"}}},"current_player":{"$ref":"#/components/schemas/Player"},"game_over":{"type":"boolean"},"winner":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Player"}]},"winner_line":{"type":["array","null"],"items":{"type":"array","items":false,"prefixItems":[{"type":"integer","minimum":0},{"type":"integer","minimum":0}]}}}},"ErrorResponse":{"type":"object","required":["detail"],"properties":{"detail":{"type":"string"}}},"MoveRequest":{"type":"object","required":["row","col"],"properties":{"col":{"type":"integer","minimum":0},"row":{"type":"integer","minimum":0}}},"Player":{"type":"string","enum":["X","O","None"]},"StartGameRequest":{"type":"object","required":["player_x_type","player_o_type"],"properties":{"human_player_symbol":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Player"}]},"player_o_type":{"type":"string"},"player_x_type":{"type":"string"}}}}},"tags":[{"name":"Tic Tac Toe","description":"三目並べゲームのAPI"}]}